Refactor "MatchTeam.py",Update "team_tournament_edition" table,Update "TournamentEdition.py".
Teams are connected to match by order of tournament phases where only teams which go trough the next phase can be connected to match which is pre-generated by number of teams in each tournament_edition.
In table "team_torunament_edition" ON DELETE CASCADE is added so some row in table can be deleted if tournament_edition is also deleted.